+++ checkmate-container/src/test/java/io/github/malczuuu/checkmate/container/plugin/PluginNo3.java
@@ -2,7 +2,9 @@ package io.github.malczuuu.checkmate.container.plugin;
 
 import io.github.malczuuu.checkmate.spi.ImageNamePlugin;
 import java.util.Optional;
+import org.jspecify.annotations.NullMarked;
 
+@NullMarked
 public class PluginNo3 implements ImageNamePlugin {
 
   @Override